+++ packages/web-components/fast-components/src/design-system-provider/design-system-provider.ts
@@ -8,47 +8,7 @@ import {
     CSSCustomPropertyTarget,
 } from "../custom-properties";
 import { isDesignSystemProvider } from "./is-design-system-provider";
-
-interface DesignSystemPropertyDeclarationConfig {
-    cssCustomProperty?: string | false;
-    default: any;
-}
-
-/**
- * Decorator to declare a property as a design-system property.
- * Accepts a config object with the following:
- *
- * default:
- * The default value of the property. Will be assigned when the use-defaults attribute is used.
- *
- * cssCustomProperty?:
- * An optional property to control the name of the css custom property being created.
- * If omitted, the css custom property will share a name with the decorated property name.
- * If assigned a false value, no css custom property will be created.
- */
-export function designSystemProperty<T extends DesignSystemProvider>(
-    config: DesignSystemPropertyDeclarationConfig
-): (source: T, property: string) => void {
-    const decorator = (
-        source: T,
-        prop: string,
-        config: DesignSystemPropertyDeclarationConfig
-    ) => {
-        if (!source.designSystemProperties) {
-            source.designSystemProperties = {};
-        }
-
-        source.designSystemProperties[prop] = {
-            cssCustomProperty:
-                config.cssCustomProperty === void 0 ? prop : config.cssCustomProperty,
-            default: config.default,
-        };
-    };
-
-    return (source: T, prop: string) => {
-        decorator(source, prop, config);
-    };
-}
+import { DesignSystemPropertyDeclarationConfig } from "./design-system-property";
 
 const supportsAdoptedStylesheets = "adoptedStyleSheets" in window.ShadowRoot.prototype;
